requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
aiokafka==0.10.0
kafka-python==2.0.2
dependency-injector==4.41.0
ruff==0.1.8
asynch==0.2.3